Make edit_temporary return a list instead of a generator
While I like the idea of making `edit_temporary` return a generator
to avoid reading the entire file in memory, returning a generator
makes cleanup problematic if the caller neglects to run the returned
generator to completion.  Since I can't devise a way to force callers
to do so, revert to returning a list.